requirements.txt
numpy>=1.20
scipy>=1.7
scikit-learn>=1.0
fpdf2>=2.7.0
